android: fix confirm-slider, money-scope, overflow and locale bugs
- SlideToConfirm: recompute drag progress at drag-end from live state; a
  stale composition-time capture meant the release check never saw real
  progress. Gates all 14 confirm surfaces.
- Move three money pipelines off composition scope onto viewModelScope so
  a back-press / scroll / sheet-dismiss can't cancel mid-move and strand
  funds: chat agent-intent execution, team-stream funding, bank payout.
- Contracts: Long math end-to-end; Int ms overflow corrupted any contract
  over ~$2,147 after funds moved.
- ChatClient SSE: break the read loop when the collector is cancelled
  (was leaking the socket).
- Wire the token-bucket card to Wallet; key payroll member rows by id;
  surface cheque cash ok:false errors; pin Locale.US on the Earn MAX
  field; fix the dash sanitizer no-op; 'instant' not 'in seconds'.
- Delete dead duplicate OnboardingBankLinkView (kyc/OnboardingBankLink is
  the wired one).

+++ android/app/src/main/java/io/talise/app/feature/contracts/ContractsApi.kt
@@ -95,7 +95,7 @@ data class ContractActionBody(val action: String = "cancel")
 data class CtrStreamPrepareBody(
     val to: String,
     val totalUsd: Double,
-    val intervalMs: Int,
+    val intervalMs: Long,
     val numTranches: Int,
 )
 
@@ -108,7 +108,7 @@ data class CtrStreamRecordBody(
     val trancheMicros: String,
     val numTranches: Int,
     val startMs: Long,
-    val intervalMs: Int,
+    val intervalMs: Long,
 )
 
 @Serializable
